initial rearranging of stat tests
